Use destScope, not srcScope for subst copy
The type in the dest(apply) is one-way copied from the type of the
func, along with all the scope inducers in it. Since the copy is
one-way, there's no way a dest GetParameter could ever escape the
scope of the source, therefore the dest scope is legitimate to use in
dest context.

This is necessary because the subst-arg in dest context should be
allowed to use the dest scope.