Oops: The hole result widget shouldn't really be displayed as in focus.
It should only emulate being in focus (post-selection) in its event map.

So we create a hidden hole result widget for its event map -- but only
for the active/selected hole result widget.

NOTE: This mechanism is more general than the current set of hacks that
allow keys being pressed on a hole result to pick it and jump to the
hole, and should replace those uglier mechanisms.