Big Revert of attempted GUI framework changes
The first change in the serie was 7751d20 at 2016 Jun 21,
some of the changes since then are good clean-ups but several changes related to that change.
The change was changing the Widget data-types to make it known whether a widget is focused before laying it out (for ExpressionGui).
It turned out to be way too much of a mess and we've come up with a better solution:

The new TreeLayout/ExpressionGui:
* Will contain its "wide" layout mode to avoid recomputing it.
* One can check if it is focused without causing re-computation, though one could also derive whether it should be focused before creating it

After the new TreeLayout implementation we should revamp names in GUI framework, document it, and then extract it to its own project.