Initial mouse clicks support. Currently just moves cursor.
The changes in Bottle.Main should perhaps move to EventMap.lookup,
but that would require to move widget's enter to its EventMap, where it probably belongs.
After that, we may add other events to respond to clicks i.e for buttons.

Though maybe first we should remove FlyNav and clean up a little..