Sugar.Types: add DefinitionOutdatedType
Towards making definitions infer with the called definitions types saved in exprUsedDefinitions.
When a used definition's type has changed, it will be marked as such
and an explicit user-action will convert to using the updated type.
For example the action could trigger adding a new function argument to a call.