Re-add hole functionality
Few things have changed (unless unknown bugs introduced ofcourse):
* The hole cursor goes to the first result automatically
* When cursor is explicitly on search term, only editing search term and selecting the first result
  are the offered options. Also in this state no hole pickers are chosen (pressing cmd+"," to add field to outer record will not pick the result).

Not sure if these changes harm the experience in any way, but the code is cleaner.

Still now it can be split to modules - HoleEdit.Open is more than 500 lines and a lot can be moved to modules such as HoleEdit.ResultWidget.
And probably a lot can still be made cleaner.
Biggest benefit of this change is that now the hole result widget isn't generated twice to take an eventmap from a "fake" widget.