Revert prefix-based animation groups (reverts de3bfd7a7)
Those were added (in 2012!) when our TextEdit animated character edits and we wanted a way to
switch between single character entities and the whole text.
That behavior has since changed and currently this is mostly a source for animation bugs.
Things seem to look nicer without it.

(For example typing euler1's "=" had an animation glitch)