"Dotter"
Due to dot having several meanings, and get-field and operators
may fragment at different places due to precedence rules,
a dot first results in a "dotter" which is still ambigiouous
and may fragment in different places depending on the next character entered.

Displayed as a "?" rather than "." to avoid display ambiguity with nullary injects.

TODO: Dotter should be available depending on type information
No point for dotter for "negate 5" which is obviously a number,
and so doesn't have a get-field option.