Add a test to the JSON codec migration.
It has a whole lot of build depends - perhaps it would be better to split Lamdu to a
library and an executable, and have the tests depend on the library.
Otherwise the test dependencies are mostly DRY.

Note that we did switch to one cabal target because we wanted to evaluate tools like
Haskell-On-Mac, intero, and other IDE plugins which seemed to rely on that.
Our current dev setups we settled on don't necessasitate iiuc.