Fix failing test, partially fix #373
https://trello.com/c/1kP4By8j/373-re-ordering-let-items-results-in-inference-error

This was broken since tag reform at 001a67dba2d0f75f1abdb965a54a56e5b8f9380a

Now when converting a lambda to record params, add the new parameter with the same tag it had,
which is also a desirable feature.

Note where this does not work: When that is also the tag for the lambda's existing parameter.

Also note that this is not consistent to when the floated lambda is already a params record -
in that case the new param does get a new tag with no name..