Cleanups left-over from annotations reform
Redundant computation of unused plShowAnnotation remained.
This were not used since 77e41d73210e3d9bf52146e0907a7525ae6d57f4
This also lead to a bug where plShowAnnotation was modified for expressions in fragments in hole results,
and that was implemented via changing the unused plShowAnnotation that had no effect.
As it actually happens to look better without this behavior I'm not fixing this "regression" for now.