optimize consistent-navigation test
The test "clicked" on many different places in the widget,
and for each one verified that moving with the cursor to some direction and back indeed goes back.
Now we only do these verifications for unique click destinations -
i.e when "clicks" at different positions result in the same destination we only do the check after the first.