Attempt at fixing build with stack on Arch Linux
See https://github.com/lamdu/lamdu/issues/104

The guess is that importing Lamdu.Prelude in a package using TH,
loads the "lamdu" target's dylib, which loads all of its dependencies too,
and that stack's GHC in Arch using too much memory crashes the build.
The change is to avoid using the lamdu target dependency in this module.