Remove the old "published tags" mechanism.
Before the "Tag Reform" (2018.02.01 #001a67dba2d0),
adding new fields to records would first add a temporary "unpublished" tag with an empty name.
These "unpublished" tags would not be offered as completions (no point having fields without names),
and if given a name (and not choosing a different tag) it would be added to the list of published tags.
(the "publish" mechanism is from around 2017.09.13 #0e2b38d0c1c8)