remove an old assertion rather than maintaining it
originally added in 2013.05.22 d4067b94cb1 -
not as an assertion but as an actual sugar condition!

then changed to an assertion in 2015.05.04 c6b0d4f49d

with fields now spread across relayed and normal, the assertion required maintanence.
but I don't think it provides much value now that everything is more complete and also much better tested!