First of toggleable sugars: Fragment
Sugar's config will have a list of which projectional syntactical sugars are enabled.
This will be useful for educational purposes and for the development process itself.

Note: seems like the config watcher doesn't respond to changes in config so currently a restart of Lamdu is required for config change to take effect.