Hyper.Ann changed - affects many places!
Ann is now more powerful and replaces previous ITerm and BTerm hypertypes too,
but in most previous cases there is now a Const wrapper.
This is a mechanic change all over the place.
It could be that some places may be generalized to not need the Const wrapper.
It makes things a bit less clean but Ann is more powerful so other things are and will be cleaner for it!