lamdu-calculus infer results don't include infer scopes
this change is towards cleanups of allowing to have a richer heterogenuous underlying AST,
with generic functions to process it.

the non-generic behavior of resolving only the types but not the scope will be replaced by resolving all.
currently the change makes things uglier as code needs to pass along the infer context
to fetch a lambda's type as a unification term to fetch the unification variable of its argument.
but it will become cleaner after later changes.