Sugar annotations parameterized further towards caching before deciding on annotation type
Instead of parameterizing the type of evaluation annotation,
sugar annotations are now wholly parameterized.

Later this annotation can be changed back to EvalPrep,
so that sugar will be cached in a manner that is agnostic to view mode.
And the annotation will be converted to a specific mode in a separate step.