Use constraints wildcards to reduce boilerplate across GUI code
This replaces long type contexts with wildcards,
reducing ctxs and imports boilerplate significantly (literally 1.5K lines),
and making things overall easier to read and write,
with the downside being that constraints may propagate freely
all to way up to the top-level consumer side.
This also replaces some instance of trying to use ConstraintKinds to alleviate the same issue.

To me this seems like a nice solution to this mtl-gone-wild problem,
and the feature has been available since GHC 7.10 (2015), see
https://www.schoolofhaskell.com/user/thomasw/new-in-ghc-7-10-partial-type-signatures#extra-constraints-wildcard

On the way of this big mechanic change have also did a little Momentu imports cleanup
(insignificant in the stats).