Re-revert "HoleEdit.ResultGroups are fuzzy matched", fixes failing test
Temporarily remove fuzzy-search functionality in holes,
fixing the sort order of results to prefer prefix-matching locals to prefix-matching globals.
Fuzzy search can be re-applied in the future,
but it should be in a way that doesn't lose the locals over globals order,
and now there's a test to verify this behaviour.

This reverts commit b6dcbb058ebb6200b39f40b0471da90e5f2d8cf0.