Move some types from Sugar.Types.Expression to Parts
The design rule is that only the types included in the corecursive group,
that is, those that cannot be extracted a separate module, are together in the Expression module.

The other types are in other modules.
This will better allow us keep the module smaller,
and enable possibly reducing incremental compile times.