Fix parens in right-to-left keyboard mode
The fix in 5bb8d5f2441a72788 was incorrect,
it did make the RTL open paren right when switching to Hebrew but not switching
the OS keyboard mode to Hebrew as well.
When the OS keyboard mode is switched it actually sends a "[" when the English "]" is pressed,
so we flipped what the OS flipped.

One regression: search term for "{" now appears flipped when in Hebrew.