Temporarily removing type annotations, towards implementing type-edits.
Sometimes maintaining huge changes is difficult to me,
and I find it easier to remove a functionality temporarily and readding it after (often based on its original code),
after I've implemented the parts that I was stuck on.

Might not be the most efficient approach but works for me at least motivationaly.