Work in progress: Scheme type parameterized on whether it is editable.
The "action/output" monad is Unit so things aren't editable.
The info monad is still there even though there isn't much point in asking for edit options if we can't edit.
But this implementation is simpler than also disabling that one and disabling the actual action should suffice.
Perhaps there will be info actions so then it makes sense not to disable those.