Extract classes out of GuiM (prep towards type editing UI), generalized TagEdit and GUI.TypeParams
It will be used also for tags in types, such as scheme parameter names,
where those will have two flavors: inferred and editable.

For the read-only flavor we would want the functions such as makeParamTag to produce a `TextWidget Unit`,
rather than a TextWidget in an actual actions monad.

To do so we'll want to not use GuiM but rather replace it with constraints,
which a new monad would also satisfy but for creating Unit variants of the actions.